+++ lib/src/shared_map_isolate.dart
@@ -45,8 +45,10 @@ class SharedStoreIsolateServer extends SharedStoreIsolate {
       var clientPort = m[0] as SendPort;
       var messageID = m[1] as int;
       var id = m[2] as String;
+      var callCasted = m[3] as _CallCasted;
 
-      var sharedMap = _sharedMaps[id];
+      var sharedMap =
+          callCasted(<K1, V1>() => getSharedMap<K1, V1>(id)) as SharedMap?;
 
       clientPort.send([messageID, sharedMap?.sharedReference()]);
       return;
@@ -56,13 +58,17 @@ class SharedStoreIsolateServer extends SharedStoreIsolate {
   }
 
   @override
-  FutureOr<SharedMap<K, V>?> getSharedMap<K, V>(
+  SharedMap<K, V>? getSharedMap<K, V>(
     String id, {
     SharedMapEntryCallback<K, V>? onPut,
     SharedMapEntryCallback<K, V>? onRemove,
   }) {
     var sharedMap = _sharedMaps[id];
     if (sharedMap != null) {
+      if (sharedMap is! SharedMap<K, V>) {
+        throw StateError(
+            "[$this] Can't cast `sharedMap` to `SharedMap<$K, $V>`: $sharedMap");
+      }
       return sharedMap as SharedMap<K, V>;
     }
 
@@ -81,6 +87,8 @@ class SharedStoreIsolateServer extends SharedStoreIsolate {
   String toString() =>
       'SharedStoreIsolateServer[$id]{sharedMaps: ${_sharedMaps.length}}';
 }
+
+typedef _CallCasted<K, V> = Object? Function(Object? Function<K1, V1>() f);
 
 class SharedStoreIsolateClient extends SharedStoreIsolate {
   final SendPort _serverPort;
@@ -125,14 +133,24 @@ class SharedStoreIsolateClient extends SharedStoreIsolate {
     var msgID = ++_msgIDCounter;
     var completer = _waitingResponse[msgID] = Completer();
 
-    _serverPort.send([_receivePort.sendPort, msgID, id]);
+    _CallCasted<K, V> callCasted = _buildCallCasted<K, V>();
+
+    _serverPort.send([_receivePort.sendPort, msgID, id, callCasted]);
 
     return completer.future.then((ref) {
-      var o = createSharedMap<K, V>(sharedReference: ref!);
+      if (ref == null) {
+        throw StateError(
+            "Can't get `SharedMap` `SendPort` from \"server\" instance: $id");
+      }
+      var o = createSharedMap<K, V>(sharedReference: ref);
       o.setCallbacksDynamic<K, V>(onPut: onPut, onRemove: onRemove);
       return o;
     });
   }
+
+  /// Generates the lambda [Function] that will be passed to the
+  /// [SharedMapIsolateServer] to allow the correct casted call to [getSharedMap]`<K,V>()`.
+  _CallCasted<K, V> _buildCallCasted<K, V>() => (f) => f<K, V>();
 
   @override
   SharedStoreReferenceIsolate sharedReference() =>
@@ -356,7 +374,7 @@ class SharedMapIsolateServer<K, V> extends SharedMapIsolate<K, V>
 
   @override
   String toString() =>
-      'SharedMapIsolateServer[$id@${sharedStore.id}]{entries: ${_entries.length}';
+      'SharedMapIsolateServer<$K,$V>[$id@${sharedStore.id}]{entries: ${_entries.length}}';
 }
 
 class SharedMapIsolateClient<K, V> extends SharedMapIsolate<K, V>
@@ -518,7 +536,7 @@ class SharedMapIsolateClient<K, V> extends SharedMapIsolate<K, V>
       SharedMapReferenceIsolate(id, sharedStore.sharedReference(), _serverPort);
 
   @override
-  String toString() => 'SharedMapIsolateClient[$id@${sharedStore.id}]';
+  String toString() => 'SharedMapIsolateClient<$K,$V>[$id@${sharedStore.id}]';
 }
 
 class SharedStoreReferenceIsolate extends SharedStoreReference {